working wersion to fix some margin
