--- src/components/Category/Category.scss
@@ -1,3 +1,0 @@
-@import '../../utils/fonts.scss';
-@import '../../utils/variables.scss';
-@import '../../utils/variables.scss';
