--- src/components/Category/Category.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import './Category.scss';
-
-const Category: React.FC = () => {
-  return (
-    <>
-      <h2>Shop by category</h2>
-    </>
-  );
-};
-
-export default Category;
