+++ src/components/HomePage/Carousel/Carousel.scss
@@ -1,4 +1,4 @@
-@import '../../utils/variables.scss';
+@import '../../../utils/variables.scss';
 
 .Carousel {
   display: flex;
