+++ src/components/HomePage/Categories/Categories.scss
@@ -1,0 +1,3 @@
+@import '../../../utils/fonts.scss';
+@import '../../../utils/variables.scss';
+@import '../../../utils/mixins.scss';
