+++ src/components/HomePage/Categories/Categories.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import './Categories.scss';
+
+const Categories: React.FC = () => {
+  return <h2>Shop by category</h2>;
+};
+
+export default Categories;
