+++ src/components/HomePage/Home.scss
@@ -1,0 +1,128 @@
+@import '../../utils/fonts.scss';
+@import '../../utils/mixins.scss';
+@import '../../utils/variables.scss';
+@import '../../utils/reset.scss';
+
+.homePage {
+  display: grid;
+
+  @include for-phone() {
+    @include grid-phone();
+  }
+
+  @include for-tablet() {
+    @include grid-tablet();
+    margin: 0 24px;
+  }
+
+  @include for-desktop() {
+    @include desktop-grid();
+    margin: 0 32px;
+  }
+}
+
+.title {
+  grid-area: 1 / 1 / 2 / -1;
+  padding: 0;
+
+  @include for-phone() {
+    margin: 24px 16px 0;
+    font-size: 32px;
+
+    & h1 {
+      font-size: 32px;
+      font-weight: 800;
+      font-family: Mont;
+      line-height: 41px;
+      margin: 0;
+      color: $color-primary;
+    }
+  }
+
+  @include for-tablet() {
+    margin-top: 32px;
+    margin-bottom: 0;
+
+    & h1 {
+      font-size: 48px;
+    }
+  }
+
+  @include for-desktop() {
+    margin-top: 56px;
+    & h1 {
+      font-size: 48px;
+    }
+  }
+}
+
+.carousel {
+  grid-area: 2 / 1 / 3 / -1;
+
+  @include for-phone() {
+    margin-top: 24px;
+  }
+
+  @include for-tablet() {
+    margin-top: 32px;
+  }
+
+  @include for-desktop() {
+    margin-top: 56px;
+  }
+}
+
+.newProducts {
+  grid-area: 3 / 1 / 4 / -1;
+
+  @include for-phone() {
+    margin-top: 56px;
+    margin-left: 16px;
+  }
+
+  @include for-tablet() {
+    margin-top: 64px;
+  }
+
+  @include for-desktop() {
+    margin-top: 80px;
+  }
+}
+
+.category {
+  grid-area: 4 / 1 / 5 / -1;
+
+  @include for-phone() {
+    margin-top: 56px;
+    margin-left: 16px;
+    margin-right: 16px;
+  }
+
+  @include for-tablet() {
+    margin-top: 64px;
+  }
+
+  @include for-desktop() {
+    margin-top: 80px;
+  }
+}
+
+.hotPrices {
+  grid-area: 5 / 1 / 6 / -1;
+
+  @include for-phone() {
+    margin-top: 56px;
+    margin-left: 16px;
+    margin-bottom: 64px;
+  }
+
+  @include for-tablet() {
+    margin-top: 64px;
+    margin-bottom: 64px;
+  }
+
+  @include for-desktop() {
+    margin-top: 80px;
+    margin-bottom: 80px;
+  }
+}
