+++ src/components/HomePage/HotPrices/HotPrices.scss
@@ -1,0 +1,3 @@
+@import '../../../utils/fonts.scss';
+@import '../../../utils/variables.scss';
+@import '../../../utils/variables.scss';
