+++ src/components/HomePage/NewProducts/NewProducts.scss
@@ -1,0 +1,11 @@
+@import '../../../utils/fonts.scss';
+@import '../../../utils/variables.scss';
+@import '../../../utils/mixins.scss';
+
+h1 {
+  font-family: Mont, Roboto;
+}
+
+h2 {
+  font-family: Mont, Roboto;
+}
