--- src/components/HotPrices/HotPrices.scss
@@ -1,3 +1,0 @@
-@import '../../utils/fonts.scss';
-@import '../../utils/variables.scss';
-@import '../../utils/variables.scss';
