--- src/components/HotPrices/HotPrices.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import './HotPrices.scss';
-
-const HotPrices: React.FC = () => {
-  return (
-    <>
-      <h2>Hot prices</h2>
-    </>
-  );
-};
-
-export default HotPrices;
