--- src/components/NewProducts/NewProducts.scss
@@ -1,11 +1,0 @@
-@import '../../utils/fonts.scss';
-@import '../../utils/variables.scss';
-@import '../../utils/mixins.scss';
-
-h1 {
-  font-family: Mont, Roboto;
-}
-
-h2 {
-  font-family: Mont, Roboto;
-}
