--- src/components/NewProducts/NewProducts.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import './NewProducts.scss';
-
-const NewProducts: React.FC = () => {
-  return (
-    <>
-      <h2>Brand new models</h2>
-    </>
-  );
-};
-
-export default NewProducts;
