--- src/pages/Home.scss
@@ -1,82 +1,0 @@
-@import '../utils/fonts.scss';
-@import '../utils/mixins.scss';
-@import '../utils/variables.scss';
-@import '../utils/reset.scss';
-
-.homePage {
-  display: grid;
-  grid-column-gap: 16px;
-  grid-row-gap: 0px;
-
-  @include for-phone() {
-    grid-template-columns: repeat(4, 1fr);
-  }
-
-  @include for-tablet() {
-    grid-template-columns: repeat(12, 1fr);
-  }
-
-  @include for-desktop() {
-    grid-template-columns: repeat(24, 1fr);
-  }
-}
-
-.title {
-  grid-area: 1 / 1 / 2 / -1;
-  @include for-phone() {
-    margin: 24px 16px 0;
-    font-size: 32px;
-
-    & h1 {
-      font-size: 32px;
-      font-weight: 800;
-      font-family: Mont;
-      line-height: 41px;
-      margin: 0;
-      color: $color-primary;
-    }
-
-    @include for-tablet() {
-      & h1 {
-        font-size: 48px;
-      }
-    }
-  }
-}
-
-.carousel {
-  grid-area: 2 / 1 / 3 / -1;
-
-  @include for-phone() {
-    margin-top: 24px;
-  }
-}
-
-.newProducts {
-  grid-area: 3 / 1 / 4 / -1;
-
-  @include for-phone() {
-    margin-top: 56px;
-    margin-left: 16px;
-  }
-}
-
-.category {
-  grid-area: 4 / 1 / 5 / -1;
-
-  @include for-phone() {
-    margin-top: 56px;
-    margin-left: 16px;
-    margin-right: 16px;
-  }
-}
-
-.hotPrices {
-  grid-area: 5 / 1 / 6 / -1;
-
-  @include for-phone() {
-    margin-top: 56px;
-    margin-left: 16px;
-    margin-bottom: 64px;
-  }
-}
