+++ src/utils/variables.scss
@@ -7,6 +7,7 @@ $color-white: #ffffff;
 $color-green: #27ae60;
 $color-red: #eb5757;
 $color-shadow: #17203166;
+$color-catalog: #6d6474;
 
 @mixin cardButton {
   font-family: 'Mont';
